requirements.txt
telethon==1.33.0
python-dotenv==1.0.1